+++ src/lib/components/resizable/resizable.component.ts
@@ -5,7 +5,20 @@ import { Component, HostBinding } from '@angular/core';
 	templateUrl: './resizable.component.html',
 	styleUrls: ['./resizable.component.scss']
 })
+/**
+ * Component for creating resizable table headers.
+ * Applied to `th` elements with the `resizable` attribute.
+ *
+ * @export
+ * @class ResizableComponent
+ */
 export class ResizableComponent {
+	/**
+	 * The width of the resizable element in pixels.
+	 *
+	 * @type {(number | null)}
+	 * @memberof ResizableComponent
+	 */
 	@HostBinding('style.width.px')
 	width: number | null = null;