+++ src/lib/components/table/table.component.spec.ts
@@ -491,4 +491,38 @@ describe('TableComponent', () => {
 			expect(destroySpy).toHaveBeenCalled();
 		});
 	});
+
+	/**
+	 * Test suite for the rowClass functionality.
+	 */
+	describe('rowClass functionality', () => {
+		/**
+		 * Test case to verify that a string rowClass is applied correctly.
+		 */
+		it('should apply a string rowClass', () => {
+			const row: TableRow = { data: { id: 1, name: 'John' }, selected: false, collapsed: true };
+			component.rowClass.set('my-class');
+			const rowClass = component._getRowClass(row);
+			expect(rowClass).toBe('my-class');
+		});
+
+		/**
+		 * Test case to verify that a function rowClass is applied correctly.
+		 */
+		it('should apply a function rowClass', () => {
+			const row: TableRow = { data: { id: 1, name: 'John' }, selected: false, collapsed: true };
+			component.rowClass.set((item: any) => (item.id === 1 ? 'first-row' : 'other-row'));
+			const rowClass = component._getRowClass(row);
+			expect(rowClass).toBe('first-row');
+		});
+
+		/**
+		 * Test case to verify that nothing happens if rowClass is not provided.
+		 */
+		it('should do nothing if rowClass is not provided', () => {
+			const row: TableRow = { data: { id: 1, name: 'John' }, selected: false, collapsed: true };
+			const rowClass = component._getRowClass(row);
+			expect(rowClass).toBe('');
+		});
+	});
 });