+++ src/lib/enums/row-class.enum.ts
@@ -1,0 +1,41 @@
+/**
+ * Represents the predefined CSS classes for styling table rows and list items.
+ * These classes correspond to Bootstrap's contextual colors.
+ *
+ * @export
+ * @enum {string}
+ */
+export enum RowClass {
+	/**
+	 * For success states.
+	 */
+	SUCCESS = 'row-success',
+	/**
+	 * For informational messages.
+	 */
+	INFO = 'row-info',
+	/**
+	 * For dangerous or error states.
+	 */
+	DANGER = 'row-danger',
+	/**
+	 * For warning states.
+	 */
+	WARNING = 'row-warning',
+	/**
+	 * For primary actions or states.
+	 */
+	PRIMARY = 'row-primary',
+	/**
+	 * For secondary actions or states.
+	 */
+	SECONDARY = 'row-secondary',
+	/**
+	 * For a light background.
+	 */
+	LIGHT = 'row-light',
+	/**
+	 * For a dark background.
+	 */
+	DARK = 'row-dark'
+}